% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCmetrics-class.R
\name{addPCs}
\alias{addPCs}
\alias{addPCs,QCmetrics-method}
\title{Add PCs to a \linkS4class{QCmetrics} object}
\usage{
addPCs(qcmetrics, PCs, heatmap_variables = NULL, heatmap_variables_fct = NULL)

\S4method{addPCs}{QCmetrics}(qcmetrics, PCs, heatmap_variables = NULL, heatmap_variables_fct = NULL)
}
\arguments{
\item{qcmetrics}{\linkS4class{QCmetrics} object}

\item{PCs}{data.frame containing PCs and a `Sample_Name` column}

\item{heatmap_variables}{test for an association between PCs and these variables.}

\item{heatmap_variables_fct}{if `heatmap_variables` are specified, which of them should be treated as factors.}
}
\value{
a  \linkS4class{QCmetrics} object
}
\description{
Add PCs to a \linkS4class{QCmetrics} object
}
