% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCmetrics-class.R
\name{combineQCmetrics}
\alias{combineQCmetrics}
\alias{combineQCmetrics,QCmetrics,QCmetrics-method}
\alias{combineQCmetrics,list,missing-method}
\title{Combine two or more \linkS4class{QCmetrics} objects into one object.}
\usage{
combineQCmetrics(x, y, ...)

\S4method{combineQCmetrics}{QCmetrics,QCmetrics}(
  x,
  y,
  ...,
  force = FALSE,
  include_probe_metrics = TRUE,
  npcs = 30,
  name = NULL,
  PCs = NULL,
  heatmap_variables = NULL,
  heatmap_variables_fct = NULL
)

\S4method{combineQCmetrics}{list,missing}(
  x,
  force = FALSE,
  include_probe_metrics = TRUE,
  npcs = 30,
  name = NULL,
  PCs = NULL,
  heatmap_variables = NULL,
  heatmap_variables_fct = NULL
)
}
\arguments{
\item{x}{\linkS4class{QCmetrics} object}

\item{y}{\linkS4class{QCmetrics} object}

\item{...}{\linkS4class{QCmetrics} objects}

\item{force}{force combining when the metric-files do not have the same package version,
probe_qc_method,detection P-value/beadnr threshold or are measured using different arrays.}

\item{include_probe_metrics}{Should probe metrics be included after combining? Defaults to TRUE.}

\item{npcs}{Number of control probe PCs to retain.}

\item{name}{Name of the experiment.}

\item{PCs}{Optionally, a dataframe containing array-wide PCs to add to the combined \linkS4class{QCmetrics}object.}

\item{heatmap_variables}{test for an association between PCs and these variables.}

\item{heatmap_variables_fct}{if `heatmap_variables` are specified, which of them should be treated as factors.}
}
\value{
a \linkS4class{QCmetrics} object
}
\description{
Combine two or more \linkS4class{QCmetrics} objects into one object.
This function will first check whether the metric-files were:
(1) measured using the same array
(2) generated using the same package version
(3) Used the same probe_qc_method
(4) The same detP and beadNr thresholds were used.
By default an error will be raised when any of the above checks fail.
}
