% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCmetrics-class.R
\name{getControlPCs}
\alias{getControlPCs}
\title{Return control-probe PCs for \linkS4class{QCmetrics} objects}
\usage{
getControlPCs(qcmetrics)
}
\arguments{
\item{qcmetrics}{a \linkS4class{QCmetrics} object}
}
\value{
a data.frame containing the array-wide PCs
}
\description{
Return control-probe PCs for \linkS4class{QCmetrics} objects
}
