% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCmetrics-class.R
\name{getOutlierOverview}
\alias{getOutlierOverview}
\title{getOutlierOverview}
\usage{
getOutlierOverview(qcmetrics)
}
\arguments{
\item{qcmetrics}{QCmetrics object}
}
\description{
Get an overview of the number of samples failed on each QC metric
}
