% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCmetrics-class.R
\name{getPCs}
\alias{getPCs}
\alias{getPCs,QCmetrics-method}
\alias{getControlPCs,QCmetrics-method}
\title{Return array-wide PCs for \linkS4class{QCmetrics} objects}
\usage{
getPCs(qcmetrics)

\S4method{getPCs}{QCmetrics}(qcmetrics)

\S4method{getControlPCs}{QCmetrics}(qcmetrics)
}
\arguments{
\item{qcmetrics}{a \linkS4class{QCmetrics} object}
}
\value{
a data.frame containing the array-wide PCs
}
\description{
Return array-wide PCs for \linkS4class{QCmetrics} objects
}
