% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCmetrics-class.R
\name{getPMS}
\alias{getPMS}
\alias{getPMS,QCmetrics-method}
\title{Return poly-methylation scores (PMS) for a \linkS4class{QCmetrics} object}
\usage{
getPMS(qcmetrics)

\S4method{getPMS}{QCmetrics}(qcmetrics)
}
\arguments{
\item{qcmetrics}{a \linkS4class{QCmetrics} object}
}
\value{
a data.frame containing the poly-methylation scores
(three age predictors, sex, smoking score, alcohol score, BMI, and cell type fractions)
}
\description{
returns predicted phenotypes for \linkS4class{QCmetrics} objects
(three age predictors, sex, smoking score, alcohol score, BMI, and cell type fractions)
}
