% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCmetrics-class.R
\name{getProbeOutliers}
\alias{getProbeOutliers}
\alias{getProbeOutliers,QCmetrics-method}
\title{Get probe outliers from a \linkS4class{QCmetrics} object}
\usage{
getProbeOutliers(qcmetrics)

\S4method{getProbeOutliers}{QCmetrics}(qcmetrics)
}
\arguments{
\item{qcmetrics}{\linkS4class{QCmetrics} object}
}
\description{
Get probe outliers from a \linkS4class{QCmetrics} object
}
