% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getQCmetrics.R
\name{getQCmetrics}
\alias{getQCmetrics}
\title{Create a QC-metrics file}
\usage{
getQCmetrics(
  samplesheet,
  name = NULL,
  experimenter = NULL,
  chunk_size = 40,
  geno = NULL,
  snplinker = NULL,
  logfile = NULL,
  verbose = TRUE,
  probe_qc_method = c("post_sampleqc", "pre_sampleqc"),
  heatmap_variables = NULL,
  heatmap_variables_fct = NULL,
  sample_qc_thresholds = NULL,
  probe_qc_thresholds = NULL,
  detp_threshold = 1e-16,
  beadnr_threshold = 3,
  relatedness_thresholds = list(IBS_mean = 1.5, IBS_var = 0.35),
  custom_outliers = NULL,
  epidish_method = c("RPC", "CBS", "CP"),
  tissue = c("blood", "breast", "epithelium"),
  calculate_PCs = FALSE,
  n_probes_PCA = NULL,
  n_estimate_PCA = chunk_size,
  nrPCs = 10,
  save_all_relationships = NULL,
  path_failed_measurements = NULL,
  cores = 1,
  BPPARAM = NULL,
  force = FALSE
)
}
\arguments{
\item{samplesheet}{Samplesheet. Should contain the following columns: Sample_Name','Sex' (M/F) and 'Basename'.}

\item{name}{Optionally, provide a name for the experiment (will be included in metadata).}

\item{experimenter}{Optionally, name of the experiment (will be included in metadata).}

\item{chunk_size}{Size of chunks, default = 40. Larger chunks will be a little bit faster
but will require more RAM.}

\item{geno}{An optional matrix with genotypes (rows = snps, columns = samples)
to check for concordance between DNAm and genotype data. See the tutoriols for more information.}

\item{snplinker}{Two-column data frame that links DNAm probes to genotype probes (first column = DNAm probes, second column = genotype IDs).
If not provided the 65 designated SNP-probes will be used.}

\item{logfile}{Filepath to write logfile to, by default no logfile is written (`logfile = NULL`).}

\item{verbose}{If `TRUE` (the default), information is printed in the console.}

\item{probe_qc_method}{Should probe QC be performed after sample QC ('post_sampleqc') or prior to sample QC ('pre_sampleqc')?.
If 'post_sampleqc', sample outliers will be removed before performing site QC.}

\item{heatmap_variables}{Character vector specifying additional variables to test for association with PCs (default = `NULL`).}

\item{heatmap_variables_fct}{Relevant if the `heatmap_variables` argument is not `NULL`.
Which of the variables specified in `heatmap_variables` should be treated as factors?}

\item{sample_qc_thresholds}{List of sample QC thresholds, if `NULL` (the default),
default thresholds will be used.
Note that custom thresholds can be specified for a subset of metrics, default
thresholds will then be used for metrics that are not specified.}

\item{probe_qc_thresholds}{List of probe QC thresholds, if `NULL` (the default),
default thresholds will be used.}

\item{detp_threshold}{detection P-value threshold (default = 1e-16).}

\item{beadnr_threshold}{beadnr threshold (default = 3).}

\item{relatedness_thresholds}{Initial relatedness thresholds, can be modified in the \code{\link{QCapp}}.}

\item{custom_outliers}{Optional additional column name(s) in the samplesheet that
that indicate that samples should be treated as outliers.
The indicated columns should be of class `logical`, (`TRUE` = outlier).}

\item{epidish_method}{Which method to use in the EpiDish \link[EpiDISH]{epidish} cellcount prediction (RPC, CBS, or CP).
Default = 'RPC'.}

\item{tissue}{Tissue-type for cellcount-estimation.
Currently 'blood', 'breast' and 'epithelium' are implemented, following the \link[EpiDISH]{epidish} function.}

\item{calculate_PCs}{Should PCA be performed? (TRUE/FALSE).
Default = `FALSE`. For large sample sizes RAM could become a bottleneck.
To limit RAM usage (besides setting this parameter to `FALSE` to skip PCA), the
`n_probes_PCA` argument can be used to limit PCA to a subset of most variable probes.}

\item{n_probes_PCA}{Number of most variable probes to maintain for PCA,
defaults to 'NULL' (all probes are used).}

\item{n_estimate_PCA}{Relevant if `n_probes_PCA` is not `NULL`.
This argument specifies the number of samples used to estimate the most variable probes (defaults to 'chunk_size').}

\item{nrPCs}{nr of PCs to retain. Default = 10}

\item{save_all_relationships}{Save all relatedness data (IBS)?
Relatedness output contains N^2 rows, and thus it will rapidly increase in size with large sample sizes.
If `FALSE`, relations defined as unrelated (based on the relatedness thresholds) will not be saved.
If `NULL` (the default), the parameter will be set to `FALSE` for sample sizes >= 1000, and to `TRUE` otherwise.}

\item{path_failed_measurements}{Filepath (.rds extension) for saving a matrix that indicates for each measurement whether it failed
based on the detP and beadNr cutoffs. Defaults to `NULL`, in which case this matrix is not saved.
This option is useful for the \code{\link{mask_measurements}} function,
which sets failed measurements in a beta-matrix to missing.}

\item{cores}{Number of cores to use, defaults to 1.
The `BPPARAM` argument controls parallelization (see: \href{https://bioconductor.org/packages/devel/bioc/vignettes/BiocParallel/inst/doc/Introduction_To_BiocParallel.pdf}{BiocParallel}).}

\item{BPPARAM}{Relevant if the `cores` argument > 1.
Specifies the parallel backend to use
(see \href{https://bioconductor.org/packages/devel/bioc/vignettes/BiocParallel/inst/doc/Introduction_To_BiocParallel.pdf}{BiocParallel}),
defaults to \code{\link[BiocParallel]{MulticoreParam}}}

\item{force}{See the \link[minfi]{minfi} documentation: Should reading different size IDAT files be forced?
Defaults to FALSE.}
}
\value{
A \linkS4class{QCmetrics} object.
}
\description{
Extract QC metrics and outlier information from IDAT files.
Returns summarized QC metrics, sample and probe outliers, relatedness info and
DNA-methylation-based predictors (PMSs) of various traits/exposures
in a \linkS4class{QCmetrics} object, which can be queried interactively using the \code{\link{QCapp}} method.
}
\details{
Although many parameters can be tweaked,
only the `samplesheet` parameter is mandatory.
Note that the samplesheet must contain the following columns: Sample_Name','Sex' (M/F) and 'Basename'.
The 'Basename' column should contain paths to the IDAT files.
Optionally, 'Age', 'Smoking', 'BMI' and 'Alcohol' columns can be included.
If these are provided provided, these can be compared to DNA-methylation-based predictors (PMSs)
of the respective variables in the \code{\link{QCapp}}.
All other included columns are kept and can be used within the \code{\link{QCapp}} to color the variables by those columns
}
\examples{
## Load example data from minfiData package (6 samples)
library(minfiData)
library(dplyr)
baseDir <- system.file("extdata", package="minfiData")
samplesheet <- read.metharray.sheet(baseDir)
# Make sure that the samplesheet contains a 'Sample_Name'
# column with the sample IDs and a 'Sex' column!
samplesheet <- samplesheet \%>\%
                   dplyr::mutate(Sample_Name =  paste(samplesheet$Slide,
                                 samplesheet$Array, sep = "_")) \%>\%
                   dplyr::rename(Sex = sex)

## Get QC metrics
metrics <- getQCmetrics(samplesheet,
  name = "example minfiData",
  calculate_PCs = TRUE,
  n_probes_PCA = 20000,
  heatmap_variables = c("Array", "Slide"),
  heatmap_variables_fct = c("Array", "Slide"))

# Add some custom thresholds:
# Note that you can specify only the thresholds that you want to change,
# for the others the default values will be used
 thresholds <- list(MU = 2000, OP = 13, RG_ratio = 0.6)
 metrics <- getQCmetrics(samplesheet,
  name = "example minfiData",
  calculate_PCs = TRUE,
  n_probes_PCA = 20000,
  sample_qc_thresholds = thresholds,
  heatmap_variables = c("Array", "Slide"),
  heatmap_variables_fct = c("Array", "Slide"))
}
