% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCmetrics-class.R
\name{getRelatedPairs}
\alias{getRelatedPairs}
\alias{getRelatedPairs,QCmetrics-method}
\title{Get related pair(s) of samples from a \linkS4class{QCmetrics} object}
\usage{
getRelatedPairs(qcmetrics)

\S4method{getRelatedPairs}{QCmetrics}(qcmetrics)
}
\arguments{
\item{qcmetrics}{\linkS4class{QCmetrics} object}
}
\description{
Get related pair(s) of samples from a \linkS4class{QCmetrics} object
}
