% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCmetrics-class.R
\name{getSampleOutliers}
\alias{getSampleOutliers}
\alias{getSampleOutliers,QCmetrics-method}
\title{Get sample outliers from a \linkS4class{QCmetrics} object}
\usage{
getSampleOutliers(qcmetrics)

\S4method{getSampleOutliers}{QCmetrics}(qcmetrics)
}
\arguments{
\item{qcmetrics}{A \linkS4class{QCmetrics} object}
}
\description{
Get sample outliers from a \linkS4class{QCmetrics} object
}
