% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCmetrics-class.R
\name{getSamplesheet}
\alias{getSamplesheet}
\alias{getSamplesheet,QCmetrics-method}
\title{Return samplesheet in a \linkS4class{QCmetrics} object}
\usage{
getSamplesheet(qcmetrics)

\S4method{getSamplesheet}{QCmetrics}(qcmetrics)
}
\arguments{
\item{qcmetrics}{a \linkS4class{QCmetrics} object}
}
\value{
a data.frame containing the sample info
}
\description{
Return samplesheet in a \linkS4class{QCmetrics} object
}
