% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCmetrics-class.R
\name{getThresholds}
\alias{getThresholds}
\alias{getThresholds,QCmetrics-method}
\title{Retrieve thresholds from a \linkS4class{QCmetrics} object}
\usage{
getThresholds(qcmetrics)

\S4method{getThresholds}{QCmetrics}(qcmetrics)
}
\arguments{
\item{qcmetrics}{\linkS4class{QCmetrics} object}
}
\description{
Retrieve thresholds from a \linkS4class{QCmetrics} object
}
