% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCmetrics-class.R
\name{makeBackground}
\alias{makeBackground}
\alias{makeBackground,QCmetrics-method}
\title{Convert QCmetrics class to QCmetricsBG class, dropping al information not needed for background representation.}
\usage{
makeBackground(qcmetrics)

\S4method{makeBackground}{QCmetrics}(qcmetrics)
}
\arguments{
\item{qcmetrics}{QCmetrics object}
}
\description{
Convert QCmetrics class to QCmetricsBG class, dropping al information not needed for background representation.
}
