% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask_measurements.R
\name{mask_measurements}
\alias{mask_measurements}
\title{Mask measurements}
\usage{
mask_measurements(
  beta,
  mask_matrix,
  output = NULL,
  chunk_size = 10000,
  delim = "\\t",
  show_progress = TRUE
)
}
\arguments{
\item{beta}{beta-matrix or a filepath pointing to a txt-file where the first column contains the probe identifiers, and the remaining columns are samples.}

\item{mask_matrix}{logical matrix (TRUE/FALSE), where rows represent probes, and columns represent samples.
`TRUE` values indicate that the measurement should be masked.
Probes and samples present in the beta-matrix should also be present in this matrix.
A masking matrix can be generating by specifying the `path_failed_measurements` parameter in \code{\link{getQCmetrics}}.}

\item{output}{Ignored if `beta` is a beta-matrix, but should be specified if `beta` is a filepath.
In that case the beta-matrix will be loaded in chunks (of size `chunk_size`) and written to `output`.}

\item{chunk_size}{Relevant if `beta` is a filepath; specifies the number of probes to load at a time.
Defaults to 10,000.}

\item{delim}{Relevant if `beta` is a filepath, defaults to `\\t`.}

\item{show_progress}{Display a progress bar? (Relevant if `beta` is a filepath).}
}
\description{
Mask measurements in a beta-matrix, based on a logical matrix that indicates which measurements
should be masked.
}
