% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCmetrics-class.R
\name{metadata,QCmetrics-method}
\alias{metadata,QCmetrics-method}
\title{retrieve metadata for a \linkS4class{QCmetrics} object}
\usage{
\S4method{metadata}{QCmetrics}(x)
}
\arguments{
\item{x}{\linkS4class{QCmetrics} object}
}
\description{
Retrieve metadata for a \linkS4class{QCmetrics} object
}
