% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCmetrics-class.R
\docType{class}
\name{QCmetrics-class}
\alias{QCmetrics-class}
\alias{show,QCmetrics-method}
\alias{ncol,QCmetrics-method}
\alias{nrow,QCmetrics-method}
\alias{dim,QCmetrics-method}
\alias{colnames,QCmetrics-method}
\alias{rownames,QCmetrics-method}
\title{S4 class to represent QCmetrics}
\usage{
\S4method{show}{QCmetrics}(object)

\S4method{ncol}{QCmetrics}(x)

\S4method{nrow}{QCmetrics}(x)

\S4method{dim}{QCmetrics}(x)

\S4method{colnames}{QCmetrics}(x)

\S4method{rownames}{QCmetrics}(x)
}
\arguments{
\item{object}{a \linkS4class{QCmetrics} object}

\item{x}{An object of class \linkS4class{QCmetrics}}
}
\description{
Returned by the \code{\link{getQCmetrics}} function. The \code{\link{QCapp}} method interactively visualizes the QCmetrics object.
}
\section{Methods (by generic)}{
\itemize{
\item \code{show}: show summary info

\item \code{ncol}: number of samples

\item \code{nrow}: number of probes

\item \code{dim}: number of probes and number of samples

\item \code{colnames}: returns sample IDs

\item \code{rownames}: returns probe IDs
}}

\section{Slots}{

\describe{
\item{\code{Sample_Metrics}}{Contains all sample QC metrics}

\item{\code{Probe_Metrics}}{Contains all probe QC metrics}

\item{\code{samplesheet}}{User-provided samplesheet}

\item{\code{ibs}}{List containing IBS (identity-by-state) info.}

\item{\code{PCs}}{Array-wide principal components}

\item{\code{controlPCs}}{Control probe principal components}

\item{\code{heatmap_pvals}}{Associations between array-wide PCs and technical/biological variables}

\item{\code{heatmap_pvals_controlPCA}}{Associations between control-probe PCs and technical/biological variables}

\item{\code{cellcounts}}{data.frame containing predicted white blood cell proportions}

\item{\code{controlmatrix}}{matrix containing control probe info}

\item{\code{pms}}{data.frame containg polymethylation scores for age, smoking, alcohol and BMI}

\item{\code{probe_qc_method}}{Which probe QC method is used (before or after sample QC)}

\item{\code{sample_qc_outliers}}{Sample outliers and thresholds}

\item{\code{probe_qc_outliers}}{Probe outliers and thresholds}

\item{\code{detp_threshold}}{Detection p-value threshold used in the `getQCmetrics()` function}

\item{\code{beadnr_threshold}}{Beadnr threshold used in the `getQCmetrics()` function}

\item{\code{nprobes_PCA}}{number of probes used for array-wide PCA, either "all" or a number.}

\item{\code{relatedness_thresholds}}{thresholds used for defining related samples}

\item{\code{selected_samples}}{Samples selected within the \code{\link{QCapp}}.}

\item{\code{metadata}}{metadata}
}}

