% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCmetrics-class.R
\docType{class}
\name{QCmetricsBG-class}
\alias{QCmetricsBG-class}
\alias{show,QCmetricsBG-method}
\alias{ncol,QCmetricsBG-method}
\alias{combineQCmetrics,QCmetricsBG,QCmetricsBG-method}
\title{S4 class to represent QCmetrics background}
\usage{
\S4method{show}{QCmetricsBG}(object)

\S4method{ncol}{QCmetricsBG}(x)

\S4method{combineQCmetrics}{QCmetricsBG,QCmetricsBG}(x, y, ..., force = FALSE, name = NULL)
}
\arguments{
\item{object}{a QCmetricsBG object}

\item{x}{\code{QCmetricsBG} object}

\item{y}{\code{QCmetricsBG} object}

\item{...}{Additional arguments.}

\item{force}{force combining when the metric-files do not have the same package version, probe_qc_method or detection P-value/beadnr threshold.}

\item{name}{name}
}
\description{
Returned by the \code{\link{makeBackground}} method.

Combine QCmetricsBG objects
}
\section{Methods (by generic)}{
\itemize{
\item \code{show}: show summary info

\item \code{ncol}: number of samples

\item \code{combineQCmetrics}: number of samples
}}

\section{Slots}{

\describe{
\item{\code{Sample_Metrics}}{Contains all sample QC metrics}

\item{\code{metadata}}{metadata}
}}

