% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCmetrics-class.R
\name{relatedness}
\alias{relatedness}
\alias{relatedness,QCmetrics-method}
\title{Extract all relatedness info from a QCmetrics object}
\usage{
relatedness(qcmetrics)

\S4method{relatedness}{QCmetrics}(qcmetrics)
}
\arguments{
\item{qcmetrics}{QCmetrics object}
}
\description{
Extract all relatedness info from a QCmetrics object
}
