% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCmetrics-class.R
\name{savePlots}
\alias{savePlots}
\title{Save QC plots contained in a \linkS4class{QCmetrics} object in high quality.}
\usage{
savePlots(
  qcmetrics,
  outdir,
  name,
  background = NULL,
  background_alpha = 0.3,
  width = 7,
  height = 5,
  unit = "in",
  filetype = "png",
  dpi = 300,
  colorby = "None",
  colorbyQCmetric = "None",
  show_outliers = TRUE,
  show_which_outliers = c("all", "metric"),
  selected_samples = c(),
  probes_plot_cutoff = 0.01,
  text_size = 13,
  point_size = 2.5,
  point_size_outlier = point_size * 1.5,
  outlier_col = "red",
  alpha = 0.8,
  theme = theme_classic(),
  colorscheme = NULL
)
}
\arguments{
\item{qcmetrics}{\linkS4class{QCmetrics} object}

\item{outdir}{path to the directory where the plots should be save (will be created if it does not exist)}

\item{name}{filename}

\item{background}{A \linkS4class{QCmetrics} or \linkS4class{QCmetricsBG} object that will be used as background.
Defaults to `NULL`.}

\item{background_alpha}{transparency, defaults to 0.3.}

\item{width}{Plot size in units ("in", "cm", or "mm"). Defaults to 7.}

\item{height}{Plot size in units ("in", "cm", or "mm"). Defaults to 5.}

\item{unit}{"in", "cm", or "mm". Defaults to "in"}

\item{filetype}{Device to use ("png", "eps", "ps", "tex" (pictex), "pdf", "jpeg", "tiff", "png", "bmp", "svg" or "wmf). Defaults to "png".}

\item{dpi}{Plot resolution. Defaults to 300.}

\item{colorby}{Column to color the plot by. Defaults to "None".}

\item{colorbyQCmetric}{Color by outlier type (MU_outlier etc.)}

\item{show_outliers}{Show outliers? (TRUE/FALSE). Based on outliers as defined in the \linkS4class{QCmetrics} object.
Defaults to `TRUE`.}

\item{show_which_outliers}{"all" or "metric".
The former will highlight all outliers (based on all metrics),
the latter will only the outliers based on the metric plotted.}

\item{selected_samples}{Highlight user-provided vector of Sample Names.}

\item{probes_plot_cutoff}{Limit the number of probes plotted in the probe QC plots.
Defaults to 0.01 (only showing probes with percentage > 0.01).}

\item{text_size}{Text size, defaults to 13.}

\item{point_size}{Point size, defaults to 2.5.}

\item{point_size_outlier}{Point size for highlighted outliers, defaults to}

\item{outlier_col}{Color used for highlighting outliers, defaults to "red".}

\item{alpha}{transparency, defaults to 0.8.}

\item{theme}{ggplot theme to use, defaults to "theme_classic()"}

\item{colorscheme}{which colorscheme to use, defaults to ggplot2::scale_color_viridis_c()/ggplot2::scale_color_viridis_d()}
}
\description{
Save QC plots contained in a \linkS4class{QCmetrics} object in high quality.
}
\examples{
## Load example data from minfiData package (6 samples)
library(minfiData)
library(dplyr)
baseDir <- system.file("extdata", package="minfiData")
samplesheet <- read.metharray.sheet(baseDir)
# Make sure that the samplesheet contains a 'Sample_Name'
# column with the sample IDs and a 'Sex' column!
samplesheet <- samplesheet \%>\%
                   dplyr::mutate(Sample_Name =  paste(samplesheet$Slide,
                                 samplesheet$Array, sep = "_")) \%>\%
                   dplyr::rename(Sex = sex)

## Get QC metrics
metrics <- getQCmetrics(samplesheet,
  name = "example minfiData",
  calculate_PCs = TRUE,
  n_probes_PCA = 20000,
  heatmap_variables = c("Array", "Slide"),
  heatmap_variables_fct = c("Array", "Slide"))

## Save plots
\dontrun{
savePlots(metrics,
          outdir = "./",
          name = "example_minfidata")
}

## color by 'Slide' variable and show metric-specific outliers:
\dontrun{
savePlots(metrics,
          outdir = "./",
          colorby = "Slide",
          show_which_outliers = "metric",
          name = "example_minfidata")
}

## Change dimensions and use pdf-format
\dontrun{
savePlots(metrics,
          outdir = "./",
          colorby = "Slide",
          show_which_outliers = "metric",
          name = "example_minfidata",
          width = 10, height = 8, filetype="pdf"
          )
}
}
