% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCmetrics-class.R
\name{updateQCmetrics}
\alias{updateQCmetrics}
\title{Update a QCmetrics object with a cleaned/normalized beta-matrix}
\usage{
updateQCmetrics(
  qcmetrics,
  beta,
  samplesheet = NULL,
  beta_format = NULL,
  chunk_size = 500,
  read_chunks = TRUE,
  delimiter = "\\t",
  calculate_PCs = FALSE,
  n_probes_PCA = NULL,
  n_estimate_PCA = chunk_size,
  nrPCs = 10,
  epidish_method = c("RPC", "CBS", "CP"),
  heatmap_variables = NULL,
  heatmap_variables_fct = NULL,
  verbose = TRUE
)
}
\arguments{
\item{qcmetrics}{QCmetrics object}

\item{beta}{beta-matrix or a filepath pointing to a txt-file where the first column contains the probe identifiers, and the remaining columns are samples.}

\item{samplesheet}{samplesheet}

\item{beta_format}{if a filepath is provided to the 'beta' parameter (instead of an R matrix), this parameter describes the format ('rds' or 'txt'). This is inferred from the file extension if this parameter is empty.}

\item{chunk_size}{chunk size, applicable if 'read_chunks'=TRUE and a filepath to a txt-file is provided in the 'beta' parameter.}

\item{read_chunks}{TRUE/FALSE, read the data in chunks? (applicable if 'beta' points toward a txt-file.)}

\item{delimiter}{Defaults to `\\t`, only applicable for txt-files.}

\item{calculate_PCs}{should PCA be performed? (TRUE/FALSE). Default = TRUE}

\item{n_probes_PCA}{number of most variable probes to maintain for PCA, defaults to 'NULL' (all probes are used).}

\item{n_estimate_PCA}{number of samples used to estimate the most variable probes (defaults to 'chunk_size').}

\item{nrPCs}{nr of PCs to retain. Default = 10}

\item{epidish_method}{Which method to using in the EpiDish cellcount prediction (RPC, CBS, or CP)}

\item{heatmap_variables}{If heatmap = y, this parameter can be used to add extra variables (in a character vector) that should be tested for an association with PCs.}

\item{heatmap_variables_fct}{If extra variables are supplied for heatmap (heatmap_variables option), which of them should be treated as a factor?}

\item{verbose}{Should the function be verbose? (TRUE/FALSE)}
}
\description{
Update a QCmetrics object with a cleaned/normalized beta-matrix
}
